requirements.txt
asgiref==3.5.1
attrs==21.4.0
Django==4.0.4
djangorestframework==3.13.1
flake8==4.0.1
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
pytest==6.2.5
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
